update build system, update deps, start new module pattern
