--- .babelrc
@@ -1,7 +1,0 @@
-{
-  "presets": [
-  	["env", {
-  	  "modules": false
-  	}]
-  ]
-}
