+++ .nvmrc
@@ -1,1 +1,1 @@
-v8.9.0
+v12.14.1
