+++ build-system/config.eslist.js
@@ -1,0 +1,12 @@
+module.exports = {
+  env: {
+    browser: true,
+    es6: true
+  },
+  extends: ['standard'],
+  parserOptions: {
+    ecmaVersion: 2018,
+    sourceType: 'module'
+  },
+  rules: {}
+}
