+++ package.json
@@ -8,28 +8,49 @@
   ],
   "license": "MIT",
   "repository": "github:ufhealth/butr",
-  "main": "dist/butr.common.js",
-  "module": "dist/butr.common.js",
-  "jsnext:main": "src/butr.js",
+  "main": "dist/butr.js",
   "files": [
     "src",
     "dist/*.js"
   ],
   "scripts": {
-    "dev": "webpack-dev-server --config webpack.dev.js",
-    "build": "webpack --config webpack.prod.js"
+    "js:watch": "webpack -w --config build-system/config.webpack.js",
+    "js:build": "webpack --config build-system/config.webpack.js",
+    "js:lint": "eslint 'src/js/**/*.js' -c build-system/config.eslint.js",
+    "manifest": "node build-system/manifest.js",
+    "reload": "browser-sync start -c build-system/config.browsersync.js",
+    "dev": "NODE_ENV=development run-p js:watch reload manifest",
+    "build": "NODE_ENV=production run-s js:build manifest",
+    "bump": "shipit",
+    "post-bump": "npm install",
+    "release": "run-s \"bump {@}\" post-bump build --"
   },
-  "dependencies": {
-    "object-assign": "^4.1.1"
-  },
+  "browserslist": [
+    "> 1% in US, ie 11"
+  ],
   "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-preset-env": "^1.6.0",
-    "html-webpack-plugin": "^2.30.1",
-    "uglifyjs-webpack-plugin": "^0.4.6",
-    "webpack": "^3.6.0",
-    "webpack-dev-server": "^2.9.1",
-    "webpack-merge": "^4.1.0"
+    "classlist-polyfill": "^1.2.0",
+    "core-js": "^3.1.4",
+    "@babel/core": "^7.5.5",
+    "@babel/polyfill": "^7.7.0",
+    "@babel/preset-env": "^7.5.5",
+    "@babel/preset-react": "^7.0.0",
+    "@ufhealth/shipit": "git+ssh://git@gitlab.ahc.ufl.edu:webservices/dev-ops/shipit.git#master",
+    "babel-loader": "^8.0.6",
+    "browser-sync": "^2.26.7",
+    "eslint": "^6.1.0",
+    "eslint-config-standard": "^13.0.1",
+    "eslint-plugin-import": "^2.18.2",
+    "eslint-plugin-node": "^9.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-standard": "^4.0.0",
+    "glob": "^7.1.4",
+    "npm-run-all": "^4.1.5",
+    "object-assign": "^4.1.1",
+    "onchange": "^6.0.0",
+    "rev-hash": "^3.0.0",
+    "rimraf": "^2.6.3",
+    "webpack": "^4.38.0",
+    "webpack-cli": "^3.3.6"
   }
 }
